requirements.txt
requests
pytelegrambotapi
validators
dnspython
pydantic
sqlalchemy